VPN-7646: restructure translation flow out of build file
